+++ frontend/lib/routes/app_routes.dart
@@ -5,6 +5,7 @@ class AppRoute {
   static const home = 'home';
   static const login = 'login';
   static const register = 'register';
+  static const profile = 'profile';
 }
 
 class AppPath {
@@ -14,4 +15,5 @@ class AppPath {
   static const home = '/home';
   static const login = '/login';
   static const register = '/register';
+  static const profile = '/profile';
 }